refactor: move util function to period context for better organization
